updated user controller and made a seperate controller for authentication
